Massive OutConsoleGridView and ShowObjectTree refactor to prepare for unit testing.
Refactored `OutConsoleGridView` and `ShowObjectTree` for improved maintainability and performance:
- Introduced `OutGridViewWindow` and `ShowObjectTreeWindow` to encapsulate UI logic.
- Added `CachedMemberResult` and `CachedMemberResultElement` for better object hierarchy handling.
- Implemented `RegexTreeViewTextFilter` for regex-based filtering in tree views.

Enhanced data handling and UI:
- Replaced `DataTable` with direct `PSObjects` handling in `ApplicationData`.
- Improved grid view filtering, column width calculations, and status bar shortcuts.
- Enhanced tree view to display object hierarchies with filtering and status bar updates.

Refactored `TypeGetter`:
- Added caching for `FormatViewDefinition` lookups.
- Made `CastObjectsToTableView` static for reusability.

Updated dependencies and removed legacy code:
- Updated `Directory.Packages.props` to use `System.Management.Automation`.
- Removed redundant and outdated code from `OutConsoleGridView`.

Bug fixes and documentation:
- Fixed column width calculation issues and concurrency in `TypeGetter`.
- Improved error handling for invalid regex patterns.
- Added XML documentation for all new and modified classes.
