+++ Directory.Packages.props
@@ -2,6 +2,7 @@
   <ItemGroup>
     <PackageVersion Include="Microsoft.PowerShell.SDK" Version="7.4.7" />
     <PackageVersion Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageVersion Include="Terminal.Gui"  Version="2.0.0-develop.4636" />
+    <PackageVersion Include="System.Management.Automation" Version="7.4.7" />
+    <PackageVersion Include="Terminal.Gui" Version="2.0.0-develop.4636" />
   </ItemGroup>
 </Project>
