+++ src/Microsoft.PowerShell.OutGridView.Models/ApplicationData.cs
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Collections.Generic;
+using System.Management.Automation;
+
 namespace Microsoft.PowerShell.OutGridView.Models;
 
 /// <summary>
@@ -9,9 +12,9 @@ namespace Microsoft.PowerShell.OutGridView.Models;
 public class ApplicationData
 {
     /// <summary>
-    ///     Gets or sets the title displayed in the Out-GridView window.
+    ///     Gets or sets the PowerShell objects to display.
     /// </summary>
-    public string? Title { get; set; }
+    public List<object>? PSObjects { get; set; }
 
     /// <summary>
     ///     Gets or sets the output mode that determines how items can be selected and returned.
@@ -19,7 +22,13 @@ public class ApplicationData
     public OutputModeOption OutputMode { get; set; }
 
     /// <summary>
-    ///     Gets or sets the filter text to apply to the data table.
+    ///     Gets or sets the title displayed in the Out-GridView window.
+    /// </summary>
+    public string? Title { get; set; }
+
+
+    /// <summary>
+    ///     Gets or sets the filter text to apply to the data.
     /// </summary>
     public string? Filter { get; set; }
 
@@ -27,11 +36,6 @@ public class ApplicationData
     ///     Gets or sets a value indicating whether to use minimal UI mode.
     /// </summary>
     public bool MinUI { get; set; }
-
-    /// <summary>
-    ///     Gets or sets the data table containing the columns and rows to display.
-    /// </summary>
-    public DataTable? DataTable { get; set; }
 
     /// <summary>
     ///     Gets or sets a value indicating whether to use the .NET driver for rendering.
@@ -52,4 +56,5 @@ public class ApplicationData
     ///     Gets or sets the version of the module.
     /// </summary>
     public string? ModuleVersion { get; set; }
+
 }
